+++ Core/Source/ColumnWorkers/Cells/ThingDefCell.cs
@@ -42,9 +42,11 @@ public readonly struct ThingDefCell : IThingDefCell
                 .CutLeft(ContentSpacing)
                 .TakeRest(out Rect labelRect);
 
+            iconRect = iconRect.Centered(_icon!.Size);
+
             if (Event.current.type == EventType.Repaint)
             {
-                _icon!.Draw(iconRect);
+                _icon.Draw(iconRect);
                 Text!.Draw(labelRect, StringNoPad);
             }